+++ Backend/AuctionH/app/controllers/ReviewController.java
@@ -19,13 +19,13 @@ public class ReviewController extends Controller {
 	public Result get(Long id) {
 		try {
 			Users user = LoginController.getUser();
-			Reviews review = Reviews.find.query().where().conjunction()
-					.eq("user_id", user.id)
-					.eq("product_id", id)
-					.endJunction()
-					.findUnique();
-			
-			return ok(Json.toJson(review));
+			if(!user.admin) {
+				Reviews review = Reviews.find.byId(id);
+				
+				return ok(Json.toJson(review));
+			} else {
+				return badRequest();
+			}
 		} catch(Exception e) {
 			return badRequest();
 		}
@@ -36,9 +36,12 @@ public class ReviewController extends Controller {
 	public Result getAll() {
 		try {
 			Users user = LoginController.getUser();
-			List<Reviews> review = user.reviews;
-			
-			return ok(Json.toJson(review));
+			if(!user.admin) {
+				List<Reviews> reviews = user.reviews;
+				return ok(Json.toJson(reviews));
+			} else {
+				return badRequest();
+			}
 		} catch(Exception e) {
 			return badRequest();
 		}
@@ -48,13 +51,16 @@ public class ReviewController extends Controller {
 	@Security.Authenticated(Secured.class)
 	public Result create() {
 		try {
-			JsonNode objectNode = request().body().asJson().get("review");
 			Users user = LoginController.getUser();
-			
-			Reviews review = Json.fromJson(objectNode, Reviews.class);
-			review.createReview(user, objectNode);	
-			
-			return ok(Json.toJson(review));
+			if(!user.admin) {
+				JsonNode objectNode = request().body().asJson().get("review");
+				Reviews review = Json.fromJson(objectNode, Reviews.class);
+				review.createReview(user, objectNode);	
+				
+				return ok(Json.toJson(review));
+			} else {
+				return badRequest();
+			}
 		} catch(Exception e) {
 			return badRequest();
 		}
@@ -64,21 +70,21 @@ public class ReviewController extends Controller {
 	@Security.Authenticated(Secured.class)	
 	public Result update(Long id) {
 		try {
-			JsonNode jsonNode = request().body().asJson();
+			Users user = LoginController.getUser();
+			if(!user.admin) {
+				JsonNode objectNode = request().body().asJson().get("review");	
+				Reviews review = Reviews.find.byId(id);
 				
-			Reviews review = Reviews.find.query().where().conjunction()
-					.eq("user_id", jsonNode.findPath("user_id").asLong())
-					.eq("product_id", jsonNode.findPath("product_id").asLong())
-					.endJunction()
-					.findUnique();
-			
-			if(review != null) {
-				review = Json.fromJson(jsonNode, Reviews.class);
-				review.update();
-				return ok(Json.toJson(review));
+				if(review != null) {
+					review = Json.fromJson(objectNode, Reviews.class);
+					review.update();
+					return ok(Json.toJson(review));
+				} else {
+					return badRequest();
+				}
 			} else {
 				return badRequest();
-			}		
+			}
 		} catch(Exception e) {
 			return badRequest();
 		}
@@ -88,19 +94,16 @@ public class ReviewController extends Controller {
 	@Security.Authenticated(Secured.class)	
 	public Result delete(Long id) {
 		try {
-			JsonNode jsonNode = request().body().asJson();
-			
-			
-			Reviews review = Reviews.find.query().where().conjunction()
-					.eq("user_id", jsonNode.findPath("user_id").asLong())
-					.eq("product_id", jsonNode.findPath("product_id").asLong())
-					.endJunction()
-					.findUnique();
-			
-			if(review != null) {
-				review = Json.fromJson(jsonNode, Reviews.class);
-				review.delete();
-				return ok(Json.toJson(""));
+			Users user = LoginController.getUser();
+			if(!user.admin) {	
+				Reviews review = Reviews.find.byId(id);
+				
+				if(review != null) {
+					review.delete();
+					return ok(Json.toJson(""));
+				} else {
+					return badRequest();
+				}
 			} else {
 				return badRequest();
 			}
