+++ Backend/AuctionH/app/controllers/WishlistController.java
@@ -21,13 +21,12 @@ public class WishlistController extends Controller{
 	public Result get(Long id) {
 		try {
 			Users user = LoginController.getUser();
-			Wishlists wishlist = Wishlists.find.query().where().conjunction()
-					.eq("user_id", user.id)
-					.eq("product_id", id)
-					.endJunction()
-					.findUnique();
-			
-			return ok(Json.toJson(wishlist));
+			if(!user.admin) {
+				Wishlists wishlist = Wishlists.find.byId(id);
+				return ok(Json.toJson(wishlist));
+			} else {
+				return badRequest();
+			}
 		} catch(Exception e) {
 			return badRequest();
 		}
@@ -38,8 +37,12 @@ public class WishlistController extends Controller{
 	public Result getAll() {
 		try {
 			Users user = LoginController.getUser();
-			List<Wishlists> wishlist = user.wishlists;
-			return ok(Json.toJson(wishlist));
+			if(!user.admin) {
+				List<Wishlists> wishlist = user.wishlists;
+				return ok(Json.toJson(wishlist));
+			} else {
+				return badRequest();
+			}
 		} catch(Exception e) {
 			return badRequest();
 		}
@@ -49,21 +52,24 @@ public class WishlistController extends Controller{
 	@Security.Authenticated(Secured.class)
 	public Result create() {
 		try {
-			JsonNode objectNode = request().body().asJson().get("wishlist");
 			Users user = LoginController.getUser();
-			
-			Wishlists wishlistChecker = Wishlists.find.query().where().conjunction()
-					.eq("user_id", user.id)
-					.eq("product_id", objectNode.findPath("product_id").asLong())
-					.endJunction()
-					.findUnique();
-			
-			if(wishlistChecker != null) {
-				return badRequest();
+			if(!user.admin) {
+				JsonNode objectNode = request().body().asJson().get("wishlist");
+				Wishlists wishlistChecker = Wishlists.find.query().where().conjunction()
+						.eq("user_id", user.id)
+						.eq("product_id", objectNode.findPath("product_id").asLong())
+						.endJunction()
+						.findUnique();
+				
+				if(wishlistChecker != null) {
+					return badRequest();
+				} else {
+					wishlistChecker = Json.fromJson(objectNode, Wishlists.class);
+					wishlistChecker.createWishlist(user, objectNode);
+					return ok(Json.toJson(wishlistChecker));
+				}
 			} else {
-				wishlistChecker = Json.fromJson(objectNode, Wishlists.class);
-				wishlistChecker.createWishlist(user, objectNode);
-				return ok(Json.toJson(wishlistChecker));
+				return badRequest();
 			}
 		} catch(Exception e) {
 			return badRequest();
@@ -74,22 +80,21 @@ public class WishlistController extends Controller{
 	@Security.Authenticated(Secured.class)
 	public Result update(Long id) {
 		try {
-			JsonNode jsonNode = request().body().asJson();
-			
-			Wishlists wishlistChecker = Wishlists.find.query().where().conjunction()
-					.eq("user_id", jsonNode.findPath("user_id").asLong())
-					.eq("product_id", jsonNode.findPath("product_id").asLong())
-					.endJunction()
-					.findUnique();
-			
-			if(wishlistChecker != null) {
-				wishlistChecker = Json.fromJson(jsonNode, Wishlists.class);
-				wishlistChecker.update();
-				return ok(Json.toJson(wishlistChecker));
+			Users user = LoginController.getUser();
+			if(!user.admin) {
+				Wishlists wishlistChecker = Wishlists.find.byId(id);
+				JsonNode objectNode = request().body().asJson().get("wishlist");
+				
+				if(wishlistChecker != null) {
+					wishlistChecker = Json.fromJson(objectNode, Wishlists.class);
+					wishlistChecker.update();
+					return ok(Json.toJson(wishlistChecker));
+				} else {
+					return badRequest();
+				}
 			} else {
 				return badRequest();
 			}
-			
 		} catch(Exception e) {
 			return badRequest();
 		}
@@ -100,19 +105,18 @@ public class WishlistController extends Controller{
 	public Result delete(Long id) {
 		try {
 			Users user = LoginController.getUser();
-			Wishlists wishlistChecker = Wishlists.find.query().where().conjunction()
-					.eq("user_id", user.id)
-					.eq("id", id)
-					.endJunction()
-					.findUnique();	
-		
-			if(wishlistChecker != null) {
-				wishlistChecker.delete();
-				return ok(Json.toJson(""));
+			if(!user.admin) {
+				Wishlists wishlistChecker = Wishlists.find.byId(id);
+				
+				if(wishlistChecker != null) {
+					wishlistChecker.delete();
+					return ok(Json.toJson(""));
+				} else {
+					return badRequest();
+				}
 			} else {
 				return badRequest();
 			}
-			
 		} catch(Exception e) {
 			return badRequest();
 		}
